+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-pre.6] - 2024-04-26
+
+### Fixed
+- Fix for context limit warning displayed when initializing the window.
+- Fix for layout problem in Unity 6.
+
+
 ## [1.0.0-pre.5] - 2024-04-22
 
 ### Changed